Add narrow version of Loudness Meter
