+++ src/plugin.hpp
@@ -13,3 +13,4 @@ extern Model *modelResonators;
 extern Model *modelBytebeat;
 extern Model *modelCognitiveShift;
 extern Model *modelLoudnessMeter;
+extern Model *modelLoudness;
